refactor some logic and renamed files and moved custom behavior to common
