+++ apps/common/behaviours/custom-save-values.js
@@ -1,0 +1,8 @@
+module.exports = superclass => class extends superclass {
+  saveValues(req, res, next) {
+    const currentFormType = req.form.values['licence-type-applyling-for'];
+    req.sessionModel.set('currentFormType', currentFormType);
+
+    return super.saveValues(req, res, next);
+  }
+};
