+++ apps/precursor-chemicals/translations/src/en/pages.json
@@ -107,6 +107,18 @@
     "sections": {
       "about-the-applicants": {
         "header": "About the applicants"
+      },
+      "about-the-licence": {
+        "header": "About the licence"
+      },
+      "evidence": {
+        "header": "Evidence"
+      },
+      "organisation": {
+        "header": "The organisation and how it operates"
+      },
+      "finalise-application": {
+        "header": "Finalise application"
       }
     },
     "fields": {
@@ -124,6 +136,78 @@
       },
       "premises-contact-details": {
         "label": "Premises contact details"
+      },
+      "responsible-officer-details": {
+        "label": "Responsible officer details"
+      },
+      "responsible-officer-dbs-information": {
+        "label": "DBS Information"
+      },
+      "responsible-officer-dbs-update-subscription": {
+        "label": "Responsible officer signed up for DBS updates"
+      },
+      "guarantor-details": {
+        "label": "Guarantor details"
+      },
+      "guarantor-dbs-information": {
+        "label": "DBS Information"
+      },
+      "is-guarantor-subscribed": {
+        "label": "Guarantor signed up for DBS updates"
+      },
+      "has-anyone-received-criminal-conviction": {
+        "label": "Any members subject to criminal convictions"
+      },
+      "invoicing-address-details": {
+        "label": "Invoicing address"
+      },
+      "invoicing-contact-details": {
+        "label": "Invoicing contact details"
+      },
+      "chemicals-used-for": {
+        "label": "Reason for chemical application"
+      },
+      "company-registration-certificate": {
+        "label": "Company registration certificate"
+      },
+      "main-customers": {
+        "label": "Main customers"
+      },
+      "main-suppliers": {
+        "label": "Main suppliers"
+      },
+      "security-measures": {
+        "label": "Security measures"
+      },
+      "how-secure-premises": {
+        "label": "How premises is secured"
+      },
+      "storage-and-handling": {
+        "label": "Storing and handling of chemicals"
+      },
+      "chemical-stock-control": {
+        "label": "Management of chemical stock control"
+      },
+      "legitimate-use": {
+        "label": "The legitimate use of end products"
+      },
+      "operating-procedures-and-auditing": {
+        "label": "SOPS AND auditing practices around use or disposal of chemicals"
+      },
+      "licence-email": {
+        "label": "Email licence will be sent to"
+      },
+      "who-is-completing-application-details": {
+        "label": "Who completed the application"
+      },
+      "is-discharge-all-licence-responsibilities": {
+        "label": "Relevant officer to discharge licence responsibilities?"
+      },
+      "explain-not-discharge-responsibilities": {
+        "label": "Why they will not discharge licence responsibilities"
+      },
+      "extra-information": {
+        "label": "Extra information"
       }
     }
   }
